Add files via upload
This repository contains Python and Julia implementations of the Kabsch algorithm for protein structural alignments. The Kabsch algorithm is a widely used method for calculating the optimal rotation matrix that aligns two sets of points or vectors which, in this case, are represented by the coordinates of protein structures. This is still a work in progress.